Implement readonly variables
Readonly annotation is implemented by allowing messages to be called
just '?'. The implementation of '?' checks for a backreference from the
object to a ConstrainedVariable object, and calls 'make_readonly' on
that. This neccesitates having such a backreference, currently, an
instance variable is used. Had to enable instance variables on W_Z3Ptr
objects for this, which makes those variables a bit slower